benchmarks/transcode/transcode/requirements.txt
ffmpeg-python
boto3